requirements.txt
streamlit
textblob
transformers
numpy
streamlit-shap
openai
python-dotenv
shap
matplotlib
ipython
transformers[torch]
transformers[tf-cpu]
transformers[flax]
torch
deepface
opencv-python-headless